requirements.txt
fastapi
numpy
pandas
armoriq-sdk
streamlit
python-dotenv
uvicorn
pytest
httpx